add PCBA test function for 8.0
1.repair the udisk test.  res/udisktester.sh
2.repair the sdcard test.  sdcard_test.c
3.change the way to test bluetooth.    bt_test.c
4.add prompt information for device without nand when it is tested the nand. nand_test.c
5.change the direction the res.sh copy the neccessary file to the recovery. res.sh
6.add hdmi mode at recovery mode. minui_pcba/graphics_drm.c
7.solve compile problems. other files

+++ minziptwrp/Zip.c
@@ -802,7 +802,7 @@ bool mzExtractZipEntryToFile(const ZipArchive *pArchive,
     const ZipEntry *pEntry, int fd)
 {
     bool ret = mzProcessZipEntryContents(pArchive, pEntry, writeProcessFunction,
-                                         (void*)fd);
+                                         (void*)(long)fd);
     if (!ret) {
         LOGE("Can't extract entry to file.\n");
         return false;
